fix: re-wrap direct date-fns exports
The functions that used to be exported from utils/index used to have
two pretty different sets of behavior.  The functions that were created
directly inside the file would only accept dates, where as the functions
that were being re-exported from date-fns would accept either a date
or a number(and then date-fns would convert them into a date).

This difference needed to be maintained for the types not to break.
Rather than roughly half of the functions in DateHelpers acceptingeither
a number or date, and the other half just a date, I decided to re-wrap
the date-helpers functions to maintain parity.