fix: rework Day to use date-helpers
These changes introduce a new wrinkle.  The problem
arises when you need to set a default prop to something
that comes from the utils adapter.  The adapter isn't available
statically, and if you used the default adapter, it would
only work for date-fns.

The solution I came up with was to use getters instead of
a default prop.  Normally, in cases that like that I would just
do get and the name of the prop, but it's complicated here
by the fact that there's already an adapter helper called
getDate, so to make things less ambiguous, I did getDateProp
and getMonthProp