fix: make formatting apply across adapters
Different adapters in date-io use different formatting strings.
For example, to get a date string of 'Sunday, March 10th 2019'
you'd need to pass 'EEEE, MMMM do yyyy' in date-fns
and 'dddd, MMMM Do YYYY' in moment.

date-io provides a generic formats interface to handle this problem.
Instead of using the format string directly, you use a generic string
like 'fullDate24h' and then date-io looks up the correct formatting
string for that specific adapter and formats it.

For formats already supplied by date-io we're good to go, but for
anything custom, we need a way to create the unique format
strings for that format, for every adapter type.

This new cloneAdapter function and the corresponding adapterMap
solve for that problem.